Set default timeout
This commit introduces a default timeout value of 2000 ms in the database configuration for new installations. The startup_db_configurations function now checks if the settings table is empty and inserts default values, including a timeout of 1500 ms, to ensure consistent behavior and prevent request failures due to missing configuration values. Moreover is adding a warining in case of timeout <200ms